topiclab-backend/pyproject.toml
[project]
name = "topiclab-backend"
version = "0.1.0"
description = "Website account/auth backend - separate from Resonnet."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
    "pillow>=10.0.0",
]

[tool.setuptools]
packages = ["app"]
